refactor(ci): split the maintainer approval helper into two tools
Each CI tool now does one thing: check_maintainer_approval.py decides the
gate, alert_maintainer_change.py renders the approver-set delta. Neither
takes a verb argument.

The login pattern is duplicated so each tool stands alone; a test asserts
the two patterns stay identical.
