--- tasks/scripts/core_approval_test.py
@@ -1,132 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Tests for tasks/scripts/core_approval.py.
-
-Run via `mise run test:core-approval`, which provides pytest through
-`uv run --with pytest`. pytest puts this file's directory on sys.path, so the
-sibling script imports directly as `core_approval`.
-"""
-
-from __future__ import annotations
-
-import json
-
-import core_approval as ca
-
-TABLE = """# Maintainers
-
-| Name | GitHub ID | Company/Organization |
-| --- | --- | --- |
-| Derek Carr | [@derekwaynecarr](https://github.com/derekwaynecarr) | Red Hat |
-| Jim Meyer | [@purp](https://github.com/purp) | NVIDIA |
-| Mrunal Patel | [@mrunalp](https://github.com/mrunalp) | Red Hat |
-"""
-
-
-def review(login: str, state: str) -> dict:
-    return {"user": {"login": login}, "state": state}
-
-
-def test_parse_maintainers_extracts_linked_logins() -> None:
-    assert ca.parse_maintainers(TABLE) == {"derekwaynecarr", "purp", "mrunalp"}
-
-
-def test_parse_maintainers_ignores_unlinked_mentions() -> None:
-    # A prose mention must not silently grant approval rights.
-    prose = TABLE + "\nThanks to [@drive-by](mailto:nobody@example.com) too.\n"
-    assert "drive-by" not in ca.parse_maintainers(prose)
-
-
-def test_parse_maintainers_returns_empty_when_table_is_reformatted() -> None:
-    assert ca.parse_maintainers("# Maintainers\n\n- derekwaynecarr\n- purp\n") == set()
-
-
-def test_decide_fails_closed_on_unparseable_list() -> None:
-    approved, description = ca.decide("# Maintainers\n", [review("purp", "APPROVED")])
-    assert not approved
-    assert "MAINTAINERS.md" in description
-
-
-def test_decide_succeeds_on_maintainer_approval() -> None:
-    approved, description = ca.decide(TABLE, [review("purp", "APPROVED")])
-    assert approved
-    assert "@purp" in description
-
-
-def test_decide_fails_on_non_maintainer_approval() -> None:
-    approved, _ = ca.decide(TABLE, [review("outsider", "APPROVED")])
-    assert not approved
-
-
-def test_decide_matches_logins_case_insensitively() -> None:
-    approved, _ = ca.decide(TABLE, [review("PuRp", "APPROVED")])
-    assert approved
-
-
-def test_comment_after_approval_does_not_revoke_it() -> None:
-    reviews = [review("purp", "APPROVED"), review("purp", "COMMENTED")]
-    approved, _ = ca.decide(TABLE, reviews)
-    assert approved
-
-
-def test_dismissed_review_revokes_approval() -> None:
-    reviews = [review("purp", "APPROVED"), review("purp", "DISMISSED")]
-    approved, _ = ca.decide(TABLE, reviews)
-    assert not approved
-
-
-def test_out_of_order_reviews_still_respect_the_latest_position() -> None:
-    # Ordering comes from the review id, not the order the caller happened
-    # to assemble the pages in.
-    reviews = [
-        {"id": 2, "user": {"login": "purp"}, "state": "DISMISSED"},
-        {"id": 1, "user": {"login": "purp"}, "state": "APPROVED"},
-    ]
-    approved, _ = ca.decide(TABLE, reviews)
-    assert not approved
-
-
-def test_changes_requested_after_approval_revokes_it() -> None:
-    reviews = [review("purp", "APPROVED"), review("purp", "CHANGES_REQUESTED")]
-    approved, _ = ca.decide(TABLE, reviews)
-    assert not approved
-
-
-def test_format_delta_names_added_and_removed_logins() -> None:
-    after = TABLE.replace(
-        "| Mrunal Patel | [@mrunalp](https://github.com/mrunalp) | Red Hat |\n",
-        "| New Person | [@newbie](https://github.com/newbie) | NVIDIA |\n",
-    )
-    body = ca.format_delta(TABLE, after)
-    assert "@newbie" in body
-    assert "@mrunalp" in body
-
-
-def test_format_delta_reports_no_change_when_only_prose_moves() -> None:
-    body = ca.format_delta(TABLE, TABLE + "\nSee also CONTRIBUTING.md.\n")
-    assert "does not change" in body
-
-
-def test_format_delta_warns_when_the_result_parses_empty() -> None:
-    body = ca.format_delta(TABLE, "# Maintainers\n\n- purp\n")
-    assert "WARNING" in body
-
-
-def _decide_exit(tmp_path, table: str, reviews: list[dict]) -> int:
-    maintainers = tmp_path / "MAINTAINERS.md"
-    maintainers.write_text(table, encoding="utf-8")
-    reviews_file = tmp_path / "reviews.json"
-    reviews_file.write_text(json.dumps(reviews), encoding="utf-8")
-    return ca.main(
-        ["decide", "--maintainers", str(maintainers), "--reviews", str(reviews_file)]
-    )
-
-
-def test_decide_exits_zero_when_a_maintainer_approved(tmp_path) -> None:
-    assert _decide_exit(tmp_path, TABLE, [review("purp", "APPROVED")]) == 0
-
-
-def test_decide_exits_nonzero_with_no_reviews(tmp_path) -> None:
-    # The exit code is the check result, so this is the gate's actual contract.
-    assert _decide_exit(tmp_path, TABLE, []) != 0
