+++ tasks/test.toml
@@ -15,7 +15,7 @@ depends = [
   "test:packaging-assets",
   "test:codex-security-release-range",
   "test:docs-website",
-  "test:core-approval",
+  "test:maintainer-approval",
 ]
 
 ["test:docs-website"]
@@ -52,9 +52,9 @@ description = "Test Codex Security release-range resolution"
 run = "uv run --no-project --with pytest pytest -o \"python_files=*_test.py\" tasks/scripts/codex_security_range_test.py"
 hide = true
 
-["test:core-approval"]
-description = "Test the maintainer approval gate helper"
-run = "uv run --no-project --with pytest pytest -o \"python_files=*_test.py\" tasks/scripts/core_approval_test.py"
+["test:maintainer-approval"]
+description = "Test the maintainer approval gate and change-alert tools"
+run = "uv run --no-project --with pytest pytest -o \"python_files=*_test.py\" tasks/scripts/check_maintainer_approval_test.py tasks/scripts/alert_maintainer_change_test.py"
 hide = true
 
 [e2e]
